feat: 🎸 Renamed Folder [PAGE TEMPLATE]]
